added test for getBreadcrumbDataCM
